refactor(ai-chat): one export per file across new tool providers
Per review feedback. Split each schema/type file that had multiple
top-level exports into one file per export:

- webhook.schema.ts -> webhook-operation.schema.ts + utils/compile-webhook-operations.util.ts
- navigation-menu-item.schema.ts -> navigation-menu-item-{scope,type}.schema.ts
- page-layout-widget.schema.ts -> 7 schema files (page-layout-type, page-layout-tab-layout-mode,
  widget-type, widget-position, widget-configuration-passthrough, page-layout-widget,
  page-layout-tab-input) + utils/derive-grid-position-for-widget.util.ts
- *-tool-dependencies.type.ts -> dependencies type stays, context type splits into
  *-tool-context.type.ts (across webhook, navigation, page-layout, workflow-run)

Inferred `z.infer<>` types that were only consumed inside helper functions
were dropped from the surface; the helpers re-derive the type locally.
