+++ packages/twenty-server/src/engine/metadata-modules/navigation-menu-item/tools/schemas/navigation-menu-item-scope.schema.ts
@@ -1,7 +1,5 @@
 import { z } from 'zod';
 
-import { NavigationMenuItemType } from 'twenty-shared/types';
-
 export const navigationMenuItemScopeSchema = z
   .enum(['workspace', 'user'])
   .describe(
@@ -9,16 +7,3 @@ export const navigationMenuItemScopeSchema = z
       "'workspace' creates a shared navigation menu item visible to everyone (requires the LAYOUTS permission). " +
       'Twenty has no separate Favorites concept — favorites are just navigation menu items with scope=user.',
   );
-
-export const navigationMenuItemTypeSchema = z.enum([
-  NavigationMenuItemType.FOLDER,
-  NavigationMenuItemType.LINK,
-  NavigationMenuItemType.OBJECT,
-  NavigationMenuItemType.VIEW,
-  NavigationMenuItemType.RECORD,
-  NavigationMenuItemType.PAGE_LAYOUT,
-]);
-
-export type NavigationMenuItemScope = z.infer<
-  typeof navigationMenuItemScopeSchema
->;
