+++ packages/twenty-server/src/engine/metadata-modules/navigation-menu-item/tools/schemas/navigation-menu-item-type.schema.ts
@@ -1,0 +1,11 @@
+import { NavigationMenuItemType } from 'twenty-shared/types';
+import { z } from 'zod';
+
+export const navigationMenuItemTypeSchema = z.enum([
+  NavigationMenuItemType.FOLDER,
+  NavigationMenuItemType.LINK,
+  NavigationMenuItemType.OBJECT,
+  NavigationMenuItemType.VIEW,
+  NavigationMenuItemType.RECORD,
+  NavigationMenuItemType.PAGE_LAYOUT,
+]);
