+++ packages/twenty-server/src/engine/metadata-modules/navigation-menu-item/tools/types/navigation-menu-item-tool-context.type.ts
@@ -1,0 +1,4 @@
+export type NavigationMenuItemToolContext = {
+  workspaceId: string;
+  userWorkspaceId?: string;
+};
