+++ packages/twenty-server/src/engine/metadata-modules/navigation-menu-item/tools/types/navigation-menu-item-tool-dependencies.type.ts
@@ -3,8 +3,3 @@ import type { NavigationMenuItemService } from 'src/engine/metadata-modules/navi
 export type NavigationMenuItemToolDependencies = {
   navigationMenuItemService: NavigationMenuItemService;
 };
-
-export type NavigationMenuItemToolContext = {
-  workspaceId: string;
-  userWorkspaceId?: string;
-};
