+++ packages/twenty-server/src/engine/metadata-modules/page-layout/tools/schemas/page-layout-tab-layout-mode.schema.ts
@@ -1,0 +1,6 @@
+import { PageLayoutTabLayoutMode } from 'twenty-shared/types';
+import { z } from 'zod';
+
+export const pageLayoutTabLayoutModeSchema = z.enum(
+  Object.values(PageLayoutTabLayoutMode) as [string, ...string[]],
+);
