+++ packages/twenty-server/src/engine/metadata-modules/page-layout/tools/schemas/page-layout-type.schema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+import { PageLayoutType } from 'src/engine/metadata-modules/page-layout/enums/page-layout-type.enum';
+
+export const pageLayoutTypeSchema = z.enum([
+  PageLayoutType.RECORD_PAGE,
+  PageLayoutType.RECORD_INDEX,
+  PageLayoutType.STANDALONE_PAGE,
+  PageLayoutType.DASHBOARD,
+]);
