+++ packages/twenty-server/src/engine/metadata-modules/page-layout/tools/types/page-layout-tool-context.type.ts
@@ -1,0 +1,3 @@
+export type PageLayoutToolContext = {
+  workspaceId: string;
+};
