+++ packages/twenty-server/src/engine/metadata-modules/page-layout/tools/types/page-layout-tool-dependencies.type.ts
@@ -7,7 +7,3 @@ export type PageLayoutToolDependencies = {
   pageLayoutTabService: PageLayoutTabService;
   pageLayoutWidgetService: PageLayoutWidgetService;
 };
-
-export type PageLayoutToolContext = {
-  workspaceId: string;
-};
