+++ packages/twenty-server/src/engine/metadata-modules/webhook/tools/types/webhook-tool-context.type.ts
@@ -1,0 +1,3 @@
+export type WebhookToolContext = {
+  workspaceId: string;
+};
