+++ packages/twenty-server/src/engine/metadata-modules/webhook/tools/types/webhook-tool-dependencies.type.ts
@@ -3,7 +3,3 @@ import type { WebhookService } from 'src/engine/metadata-modules/webhook/webhook
 export type WebhookToolDependencies = {
   webhookService: WebhookService;
 };
-
-export type WebhookToolContext = {
-  workspaceId: string;
-};
