+++ packages/twenty-server/src/modules/workflow/workflow-run-tools/types/workflow-run-tool-context.type.ts
@@ -1,0 +1,6 @@
+import type { ActorMetadata } from 'twenty-shared/types';
+
+export type WorkflowRunToolContext = {
+  workspaceId: string;
+  actorContext?: ActorMetadata;
+};
