+++ packages/twenty-server/src/modules/workflow/workflow-run-tools/types/workflow-run-tool-dependencies.type.ts
@@ -1,14 +1,7 @@
-import type { ActorMetadata } from 'twenty-shared/types';
-
 import type { GlobalWorkspaceOrmManager } from 'src/engine/twenty-orm/global-workspace-datasource/global-workspace-orm.manager';
 import type { WorkflowTriggerWorkspaceService } from 'src/modules/workflow/workflow-trigger/workspace-services/workflow-trigger.workspace-service';
 
 export type WorkflowRunToolDependencies = {
   globalWorkspaceOrmManager: GlobalWorkspaceOrmManager;
   workflowTriggerWorkspaceService: WorkflowTriggerWorkspaceService;
-};
-
-export type WorkflowRunToolContext = {
-  workspaceId: string;
-  actorContext?: ActorMetadata;
 };
